+++ Ares.Core.Grpc/Services/AresScriptingService.cs
@@ -6,9 +6,9 @@ using System.Threading.Channels;
 using Ares.Core.Scripting;
 using Microsoft.Extensions.Logging;
 using Google.Protobuf.WellKnownTypes;
-using AresScript;
 using AresScript.ScriptAnalysis;
 using Ares.Datamodel.Scripting;
+using AresScript.Interpreters;
 
 namespace Ares.Core.Grpc.Services;
