+++ AresScript/ScriptAnalysis/AresScriptAnalysis.Completions.cs
@@ -4,7 +4,7 @@ using Ares.Datamodel.Scripting;
 using AresScript.Generated;
 using System.Text;
 using System.Text.RegularExpressions;
-using AresScript;
+using AresScript.Interpreters;
 
 namespace AresScript.ScriptAnalysis;
