+++ UI/JsInterops/MonacoHoverProvider.cs
@@ -10,6 +10,7 @@ using Ares.Datamodel.Scripting;
 using Ares.Services;
 using AresScript;
 using AresScript.Generated;
+using AresScript.Interpreters;
 using AresScript.ScriptAnalysis;
 using Google.Protobuf.WellKnownTypes;
 using Microsoft.JSInterop;
